+++ src/builtins/filters/number.rs
@@ -1,8 +1,8 @@
 /// Filters operating on numbers
 use std::collections::HashMap;
 
-use serde_json::value::{to_value, Value};
 use humansize::{file_size_opts, FileSize};
+use serde_json::value::{to_value, Value};
 
 use errors::Result;
 
@@ -36,11 +36,7 @@ pub fn round(value: Value, args: HashMap<String, Value>) -> Result<Value> {
         Some(val) => try_get_value!("round", "precision", i32, val),
         None => 0,
     };
-    let multiplier = if precision == 0 {
-        1.0
-    } else {
-        10.0_f64.powi(precision)
-    };
+    let multiplier = if precision == 0 { 1.0 } else { 10.0_f64.powi(precision) };
 
     match method.as_ref() {
         "common" => Ok(to_value((multiplier * num).round() / multiplier).unwrap()),
@@ -59,20 +55,16 @@ pub fn filesizeformat(value: Value, _: HashMap<String, Value>) -> Result<Value> 
     let num = try_get_value!("filesizeformat", "value", i64, value);
     num.file_size(file_size_opts::CONVENTIONAL)
         .or_else(|_| {
-            Err(format!(
-                "Filter `filesizeformat` was called on a negative number: {}",
-                num
-            ).into())
-        })
-        .map(to_value)
+            Err(format!("Filter `filesizeformat` was called on a negative number: {}", num).into())
+        }).map(to_value)
         .map(|x| x.unwrap())
 }
 
 #[cfg(test)]
 mod tests {
-    use std::collections::HashMap;
-    use serde_json::value::to_value;
     use super::*;
+    use serde_json::value::to_value;
+    use std::collections::HashMap;
 
     #[test]
     fn test_pluralize_single() {