+++ src/parser/tests/lexer.rs
@@ -143,7 +143,6 @@ fn lex_array() {
     }
 }
 
-
 #[test]
 fn lex_basic_expr() {
     let inputs = vec![
@@ -383,10 +382,8 @@ fn lex_include_tag() {
 #[test]
 fn lex_import_macro_tag() {
     assert!(
-        TeraParser::parse(
-            Rule::import_macro_tag,
-            "{% import \"macros.html\" as macros %}",
-        ).is_ok()
+        TeraParser::parse(Rule::import_macro_tag, "{% import \"macros.html\" as macros %}",)
+            .is_ok()
     );
 }
 
@@ -410,11 +407,8 @@ fn lex_block_tag() {
 
 #[test]
 fn lex_macro_tag() {
-    let inputs = vec![
-        "{%- macro tag() %}",
-        "{% macro my_block(name) -%}",
-        "{% macro my_block(name=42) %}",
-    ];
+    let inputs =
+        vec!["{%- macro tag() %}", "{% macro my_block(name) -%}", "{% macro my_block(name=42) %}"];
     for i in inputs {
         assert_lex_rule!(Rule::macro_tag, i);
     }
@@ -473,22 +467,12 @@ fn lex_for_tag() {
 
 #[test]
 fn lex_break_tag() {
-    assert!(
-        TeraParser::parse(
-            Rule::break_tag,
-            "{% break %}"
-        ).is_ok()
-    );
+    assert!(TeraParser::parse(Rule::break_tag, "{% break %}").is_ok());
 }
 
 #[test]
 fn lex_continue_tag() {
-    assert!(
-        TeraParser::parse(
-            Rule::continue_tag,
-            "{% continue %}"
-        ).is_ok()
-    );
+    assert!(TeraParser::parse(Rule::continue_tag, "{% continue %}").is_ok());
 }
 
 #[test]