+++ src/renderer/tests/errors.rs
@@ -1,14 +1,10 @@
-use std::collections::BTreeMap;
 use context::Context;
-use errors::Result;
 use tera::Tera;
 
 #[test]
 fn error_location_basic() {
     let mut tera = Tera::default();
-    tera.add_raw_templates(vec![
-        ("tpl", "{{ 1 + true }}"),
-    ]).unwrap();
+    tera.add_raw_templates(vec![("tpl", "{{ 1 + true }}")]).unwrap();
 
     let result = tera.render("tpl", &Context::new());
 
@@ -23,14 +19,30 @@ fn error_location_inside_macro() {
     let mut tera = Tera::default();
     tera.add_raw_templates(vec![
         ("macros", "{% macro hello()%}{{ 1 + true }}{% endmacro hello %}"),
-        ("tpl", "{% import \"macros\" as macros %}{{ macro::hello() }}"),
+        ("tpl", "{% import \"macros\" as macros %}{{ macros::hello() }}"),
     ]).unwrap();
 
     let result = tera.render("tpl", &Context::new());
 
     assert_eq!(
         result.unwrap_err().iter().nth(0).unwrap().description(),
-        "Failed to render \'tpl\': error while rendering a macro from the `macro` namespace"
+        "Failed to render \'tpl\': error while rendering macro `macros::hello`"
+    );
+}
+
+#[test]
+fn error_loading_macro_from_unloaded_namespace() {
+    let mut tera = Tera::default();
+    tera.add_raw_templates(vec![
+        ("macros", "{% macro hello()%}{{ 1 + true }}{% endmacro hello %}"),
+        ("tpl", "{% import \"macros\" as macros %}{{ macro::hello() }}"),
+    ]).unwrap();
+
+    let result = tera.render("tpl", &Context::new());
+    println!("{:#?}", result);
+    assert_eq!(
+        result.unwrap_err().iter().nth(1).unwrap().description(),
+        "Macro namespace `macro` was not found in template `tpl`. Have you maybe forgotten to import it, or misspelled it?"
     );
 }
 
@@ -71,24 +83,23 @@ fn error_location_in_parent_in_macro() {
     let mut tera = Tera::default();
     tera.add_raw_templates(vec![
         ("macros", "{% macro hello()%}{{ 1 + true }}{% endmacro hello %}"),
-        ("parent", "{% import \"macros\" as macros %}{{ macro::hello() }}{% block bob %}{% endblock bob %}"),
+        ("parent", "{% import \"macros\" as macros %}{{ macros::hello() }}{% block bob %}{% endblock bob %}"),
         ("child", "{% extends \"parent\" %}{% block bob %}{{ super() }}Hey{% endblock bob %}"),
     ]).unwrap();
 
     let result = tera.render("child", &Context::new());
+    println!("{:?}", result);
 
     assert_eq!(
         result.unwrap_err().iter().nth(0).unwrap().description(),
-        "Failed to render \'child\': error while rendering a macro from the `macro` namespace (error happened in \'parent\')."
+        "Failed to render \'child\': error while rendering macro `macros::hello` (error happened in \'parent\')."
     );
 }
 
 #[test]
 fn error_out_of_range_index() {
     let mut tera = Tera::default();
-    tera.add_raw_templates(vec![
-        ("tpl", "{{ arr[10] }}"),
-    ]).unwrap();
+    tera.add_raw_templates(vec![("tpl", "{{ arr[10] }}")]).unwrap();
     let mut context = Context::new();
     context.add("arr", &[1, 2, 3]);
 
@@ -103,9 +114,7 @@ fn error_out_of_range_index() {
 #[test]
 fn error_unknown_index_variable() {
     let mut tera = Tera::default();
-    tera.add_raw_templates(vec![
-        ("tpl", "{{ arr[a] }}"),
-    ]).unwrap();
+    tera.add_raw_templates(vec![("tpl", "{{ arr[a] }}")]).unwrap();
     let mut context = Context::new();
     context.add("arr", &[1, 2, 3]);
 
@@ -120,9 +129,7 @@ fn error_unknown_index_variable() {
 #[test]
 fn error_invalid_type_index_variable() {
     let mut tera = Tera::default();
-    tera.add_raw_templates(vec![
-        ("tpl", "{{ arr[a] }}"),
-    ]).unwrap();
+    tera.add_raw_templates(vec![("tpl", "{{ arr[a] }}")]).unwrap();
 
     let mut context = Context::new();
     context.add("arr", &[1, 2, 3]);
@@ -133,5 +140,17 @@ fn error_invalid_type_index_variable() {
     assert_eq!(
         result.unwrap_err().iter().nth(1).unwrap().description(),
         "Only variables evaluating to String or Number can be used as index (`a` of `arr[a]`)"
+    );
+}
+
+#[test]
+fn error_when_missing_macro_templates() {
+    let mut tera = Tera::default();
+    let result = tera.add_raw_templates(vec![
+        ("parent", "{% import \"macros\" as macros %}{{ macros::hello() }}{% block bob %}{% endblock bob %}"),
+    ]);
+    assert_eq!(
+        result.unwrap_err().iter().nth(0).unwrap().description(),
+        "Template `parent` loads macros from `macros` which isn\'t present in Tera"
     );
 }