+++ src/renderer/tests/square_brackets.rs
@@ -1,7 +1,6 @@
 use std::collections::HashMap;
 
 use context::Context;
-use errors::Result;
 use tera::Tera;
 
 #[derive(Serialize)]
@@ -14,10 +13,8 @@ struct Test {
 #[test]
 fn test_var_access_by_square_brackets() {
     let mut context = Context::new();
-    context.add(
-        "var",
-        &Test { a: "hi".into(), b: "i_am_actually_b".into(), c: vec!["fred".into()] },
-    );
+    context
+        .add("var", &Test { a: "hi".into(), b: "i_am_actually_b".into(), c: vec!["fred".into()] });
     context.add("zero", &0);
     context.add("a", "b");